src/glorious_agents/skills/docs/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glorious-skill-docs"
version = "0.1.0"
description = "Documentation management skill for Glorious Agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "glorious-agents>=0.1.0",
    "sqlmodel>=0.0.16",
    "sqlalchemy>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.entry-points."glorious_agents.skills"]
docs = "glorious_docs.skill:app"

[tool.hatch.build.targets.wheel]
packages = ["src/glorious_docs"]
